Changes to Race Report Service/Controller - Moved all business logic from the race report controller to the service. - All race report related logic moved from the schedule controller.
At the moment the race report service is really cluttered. I plan on taking the monolith function that all of the code for creating a race report is in and splitting it into a bunch of smaller functions that will make it much easier to read and maintain. After that im going to go and track down a bunch of small issues that arent listed. Then after that im going to get back on track and knock out some of the larger issues that I have outlined in the repository.